+++ src/views/laboratory_1/software/software/information/software-detail.vue
@@ -296,49 +296,21 @@
   //   course_id: 1,
   //   course_name: '适用课程'
   // }
-  // 下拉框信息测试
-  import { fetchSoftwareCategoryInfoById, fetchSoftwareCategoryInfos } from '@/api/laboratory_1/software'
-  import { deleteSoftwareInfoById, fetchSoftwareInfoById } from '@/api/laboratory_1/software'
 
-  import { getSelectValue } from '@/utils/get-select-value'
-  import { isChinese, isEnglish } from '@/utils/fieldValidate'
-  import DictionaryForSelect from '@/components/DictionaryForSelect'
-  import {
-    fetchApplicableSystemInfoById,
-    fetchApplicableSystemInfos
-  } from '../../../../../api/laboratory_1/applicable-system'
-  import { fetchBrandInfoById, fetchBrandInfos } from '../../../../../api/laboratory_1/brand'
-  import { fetchUnitPriceInfoById, fetchUnitPriceInfos } from '../../../../../api/laboratory_1/unit-price'
-  import { fetchSupplierInfoById, fetchSupplierInfos } from '../../../../../api/laboratory_1/supplier'
-  import { fetchManufacturerInfoById, fetchManufacturerInfos } from '../../../../../api/laboratory_1/manufacturer'
-  import { fetchCountryCodeInfoById, fetchCountryCodeInfos } from '../../../../../api/laboratory_1/country-code'
+import { getSelectValue } from '@/utils/get-select-value'
+import { fetchApplicableSystemInfoById, fetchApplicableSystemInfos } from '@/api/laboratory_1/applicable-system'
+import { fetchBrandInfoById, fetchBrandInfos } from '@/api/laboratory_1/brand'
+import { fetchUnitPriceInfoById, fetchUnitPriceInfos } from '@/api/laboratory_1/unit-price'
+import { fetchSupplierInfoById, fetchSupplierInfos } from '@/api/laboratory_1/supplier'
+import { fetchManufacturerInfoById, fetchManufacturerInfos } from '@/api/laboratory_1/manufacturer'
+import { fetchCountryCodeInfoById, fetchCountryCodeInfos } from '@/api/laboratory_1/country-code'
+import DictionaryForSelect from '@/components/DictionaryForSelect'
+import { isChinese, isEnglish } from '@/utils/fieldValidate'
+import { modifySoftwareInfo, deleteSoftwareInfoById, fetchSoftwareInfoById, } from '@/api/laboratory_1/software'
+import { fetchSoftwareCategoryInfoById, fetchSoftwareCategoryInfos } from '@/api/laboratory_1/software-category'
 
-  const fakeList = [
-  { id: 1, pre_name: 'c1', name: 'radio内容1', phone: '13512341234' },
-  { id: 2, pre_name: 'c1', name: 'radio内容2', phone: '16812341234' },
-  { id: 3, pre_name: 'c1', name: 'radio内容3', phone: '13812341234' },
-  { id: 4, pre_name: 'c1', name: 'radio内容4', phone: '13812341234' },
-  { id: 5, pre_name: 'c1', name: 'radio内容5', phone: '13812341234' },
-  { id: 6, pre_name: 'c1', name: 'radio内容6', phone: '13812341234' },
-  { id: 7, pre_name: 'c1', name: 'radio内容7', phone: '13812341234' },
-  { id: 8, pre_name: 'c1', name: 'radio内容8', phone: '13812341234' },
-  { id: 9, pre_name: 'c1', name: 'radio内容9', phone: '13812341234' },
-  { id: 10, pre_name: 'c1', name: 'radio内容10', phone: '13812341234' },
-  { id: 11, pre_name: 'c1', name: 'radio内容11', phone: '13812341234' },
-  { id: 12, pre_name: 'c2', name: 'radio内容12', phone: '13812341234' },
-  { id: 13, pre_name: 'c2', name: 'radio内容13', phone: '13812341234' },
-  { id: 14, pre_name: 'c2', name: 'radio内容14', phone: '13812341234' },
-  { id: 15, pre_name: 'c2', name: 'radio内容15', phone: '13812341234' },
-  { id: 16, pre_name: 'c2', name: 'radio内容16', phone: '13812341234' },
-  { id: 17, pre_name: 'c3', name: 'radio内容17', phone: '13812341234' },
-  { id: 18, pre_name: 'c3', name: 'radio内容18', phone: '13812341234' },
-  { id: 19, pre_name: 'c3', name: 'radio内容19', phone: '13812341234' },
-  { id: 20, pre_name: 'c3', name: 'radio内容20', phone: '13812341234' },
-  { id: 21, pre_name: 'c3', name: 'radio内容21', phone: '13812341234' },
-  { id: 22, pre_name: 'c3', name: 'radio内容22', phone: '13812341234' }
-]
 
-// 字典
+  // 字典
 const allListName = [
   { key: 'softwareCategoryList', dialogVisibleName: 'softwareCategoryDialogVisible', option: 'getSoftwareCategoryList' },
   { key: 'courseList', dialogVisibleName: 'courseDialogVisible', option: 'getCourseList' },
@@ -358,8 +330,7 @@ const key2option = [
   {key: 'supplier', option: 'fetchSupplierById'},
   {key: 'countryCode', option: 'fetchCountryCodeById'}
 ]
-
-  export default {
+export default {
   name: 'SoftwareDetail',
   components: {
     DictionaryForSelect
@@ -493,7 +464,14 @@ const key2option = [
     this.getOriginalData()
   },
   methods: {
-    /* 根据ID获取数据 */
+    /**
+      * @method：getOriginalData
+      * @desc：根据ID获取数据
+      * @params: null
+      * @create date： 2020/6/29
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async getOriginalData() {
       const id = this.$route.query.id
       console.log(id)
@@ -507,11 +485,26 @@ const key2option = [
           alert('获取内容失败！' + err)
       })
     },
-    /* 返回上一页 */
+    /**
+      * @method：handleReturn
+      * @desc：返回上一页
+      * @params: null
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     handleReturn() {
       this.$router.go(-1)
     },
     /* 获取软件分类列表信息 */
+    /**
+      * @method：getSoftwareCategoryList
+      * @desc：获取软件分类列表信息
+      * @params: keyword 查询条件，可为null
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async getSoftwareCategoryList(keyword) {
       console.log('调用获取软件分类列表信息接口')
       // 实际 keyword可为空
@@ -522,7 +515,14 @@ const key2option = [
          alert('获取用获取软件分类列表失败！' + err)
       })
     },
-    /* 获取适用系统列表信息 */
+    /**
+      * @method：getApplicableSystemList
+      * @desc：获取适用系统列表信息
+      * @params: keyword 查询条件，可为null
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async getApplicableSystemList(keyword) {
       await fetchApplicableSystemInfos(keyword)
         .then(res => {
@@ -531,7 +531,14 @@ const key2option = [
           alert('获取适用系统列表失败！' + err)
         })
     },
-    /* 获取品牌列表信息 */
+    /**
+      * @method：getBrandList
+      * @desc：获取品牌列表信息
+      * @params: keyword 查询条件，可为null
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async getBrandList(keyword) {
       await fetchBrandInfos(keyword).then(res => {
         this.brandList = res.data.list;
@@ -539,7 +546,14 @@ const key2option = [
         alert('获取品牌列表失败！' + err)
       })
     },
-    /* 获取计量单位列表信息 */
+    /**
+      * @method：getUnitList
+      * @desc：获取计量单位列表信息
+      * @params: keyword 查询条件，可为null
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async getUnitList(keyword) {
       await fetchUnitPriceInfos(keyword).then(res => {
         this.unitList = res.data.list;
@@ -547,7 +561,14 @@ const key2option = [
         alert('获取计量单位列表信息失败！' + err)
       })
     },
-    /* 获取供货商列表信息 */
+    /**
+      * @method：getSupplierList
+      * @desc：获取供货商列表信息
+      * @params: keyword 查询条件，可为null
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async getSupplierList(keyword) {
       await fetchSupplierInfos(keyword).then(res => {
         this.supplierList = res.data.list;
@@ -555,7 +576,14 @@ const key2option = [
         alert('获取供货商列表信息失败！' + err)
       })
     },
-    /* 获取生产商列表信息 */
+    /**
+      * @method：getManufacturerList
+      * @desc：获取生产商列表信息
+      * @params: keyword 查询条件，可为null
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async getManufacturerList(keyword) {
      await fetchManufacturerInfos(keyword).then(res => {
        this.manufacturerList = res.data.list;
@@ -563,7 +591,14 @@ const key2option = [
        alert('获取生产商列表信息失败！' + err)
      })
     },
-    /* 获取国码列表信息 */
+    /**
+      * @method：getCountryCodeList
+      * @desc：获取国码列表信息
+      * @params: keyword 查询条件，可为null
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async getCountryCodeList(keyword) {
       await fetchCountryCodeInfos(keyword).then(res => {
         this.countryCodeList = res.data.list;
@@ -571,8 +606,14 @@ const key2option = [
         alert('获取生产商列表信息失败！' + err)
       })
     },
-
-    /* 获取软件分类信息 id */
+    /**
+      * @method：fetchSoftwareCategoryById
+      * @desc：根据ID获取软件分类信息
+      * @params: id
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async fetchSoftwareCategoryById(id) {
       // 实际使用
       console.log('调用获取软件分类信息接口')
@@ -585,7 +626,14 @@ const key2option = [
           return null
         })
     },
-    /* 获取适用系统信息 id */
+    /**
+      * @method：fetchApplicableSystemById
+      * @desc：根据ID获取适用系统信息
+      * @params: id
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async fetchApplicableSystemById(id) {
       // 根据select id key查询出的数据 各字段经过包装
       await fetchApplicableSystemInfoById(id).then(res => {
@@ -595,7 +643,14 @@ const key2option = [
         return null
       })
     },
-    /* 获取品牌信息 id */
+    /**
+      * @method：fetchBrandById
+      * @desc：获根据ID取品牌信息
+      * @params: id
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async fetchBrandById(id) {
       // 根据select id key查询出的数据 各字段经过包装
       await fetchBrandInfoById(id).then(res => {
@@ -605,7 +660,14 @@ const key2option = [
         return null
       })
     },
-    /* 获取计量单位信息 id */
+    /**
+      * @method：fetchUnitById
+      * @desc：根据ID获取计量单位信息
+      * @params: id
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async fetchUnitById(id) {
       // 根据select id key查询出的数据 各字段经过包装
       await fetchUnitPriceInfoById(id).then(res => {
@@ -615,7 +677,14 @@ const key2option = [
         return null
       })
     },
-    /* 获取供货商信息 id */
+    /**
+      * @method：fetchSupplierById
+      * @desc：根据ID获取供货商信息
+      * @params: id
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async fetchSupplierById(id) {
       // 根据select id key查询出的数据 各字段经过包装
       await fetchSupplierInfoById(id).then(res => {
@@ -625,27 +694,49 @@ const key2option = [
         return null
       })
     },
-    /* 获取生产商信息 id */
+    /**
+      * @method：fetchManufacturerById
+      * @desc：根据ID获取生产商信息
+      * @params: id
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async fetchManufacturerById(id) {
       // 根据select id key查询出的数据 各字段经过包装
       await fetchManufacturerInfoById(id).then(res => {
         this.tempDataByID = res.data.item;
       }).catch(err => {
-        alert('获取供货商信息内容失败！' + err)
+        alert('获取生产商信息内容失败！' + err)
         return null
       })
     },
-    /* 获取国码信息 id */
+    /**
+      * @method：fetchCountryCodeById
+      * @desc：根据ID获取国码信息
+      * @params: id
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async fetchCountryCodeById(id) {
       // 根据select id key查询出的数据 各字段经过包装
       await fetchCountryCodeInfoById(id).then(res => {
         this.tempDataByID = res.data.item;
       }).catch(err => {
-        alert('获取供货商信息内容失败！' + err)
+        alert('获取国码信息内容失败！' + err)
         return null
       })
     },
     /* 利用字典1 根据关键词打开dialog */
+    /**
+      * @method：handleOpenDialog
+      * @desc：利用字典1 根据关键词打开dialog
+      * @params: dialogVisible
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async handleOpenDialog(dialogVisible) {
       this[dialogVisible] = true // 打开对应的dialog
       // 生成默认radio数据
@@ -660,42 +751,70 @@ const key2option = [
         })
       }
     },
-    /* 利用字典2 根据关键字获取对应方法 并返回查询出的数据 */
+    /**
+      * @method：getMethodByKey
+      * @desc：利用字典2 根据关键字获取对应方法 并返回查询出的数据
+      * @params: key：找出方法对关键字，id：被找出方法需要要参数值
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async getMethodByKey(key, id) {
       let option = key2option.filter(m => m.key === key).map(m => m.option).pop()
       await this[option](id)
     },
-    /* 抽取方法 将信息同步 */
+    /**
+      * @method：handleSelect
+      * @desc：抽取方法 将信息同步
+      * @params:
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     async handleSelect(data) {
-      // 查询选择的id相关的数据
+      // 查询 选择的id的相关的数据
       await this.getMethodByKey(data.methodKey, data.select_id_value)
       let tempData = this.tempDataByID
       this.tempDataByID = null
-      console.log('tempData=')
-      console.log({ ...tempData })
+      // console.log('tempData=')
+      // console.log({ ...tempData })
       getSelectValue(data, this.dataForm, tempData)
       // 表单数据验证规则
       this.$refs['dataForm'].validateField(data.select_id_key)
       // 关闭对话框
-      console.log('data.dialogVisibleKey = ' + data.dialogVisibleKey)
+      // console.log('data.dialogVisibleKey = ' + data.dialogVisibleKey)
       this[data.dialogVisibleKey] = false
     },
-    /* 筛选操作 */
+    /**
+      * @method：handleQuery
+      * @desc：筛选radio
+      * @params: data：自定义组件传的数据
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     handleQuery(data) {
-      console.log('data=>' + data.listName + ',' + data.queryKeyword)
+      // console.log('data=>' + data.listName + ',' + data.queryKeyword)
       let keyword = data.queryKeyword
       let listName = data.listName
       const option = allListName.filter(m => m.key === listName).map(m => m.option).pop()
-      console.log('option=' + option)
-      console.log('keyword=', keyword)
+      // console.log('option=' + option)
+      // console.log('keyword=', keyword)
+      // 根据option找出对应筛选方法
       if (option) {
         this[option](keyword)
       } else {
         return false
       }
     },
-
-    /* 编辑前 */
+    /**
+      * @method：beforeEdit
+      * @desc：为页面显示为编辑页面做准备
+      * @params:
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     beforeEdit() {
       this.showSaveBtn = !this.showSaveBtn
       // 保存修改前数据
@@ -707,7 +826,14 @@ const key2option = [
       // 设为不可用,防止暂存数据出错
       this.isAble = true
     },
-    /* 编辑后 */
+    /**
+      * @method：afterEdit
+      * @desc：编辑成功后的操作，还原修改前所有状态
+      * @params:
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     afterEdit() {
       // 还原修改前所有状态
       this.tempData = null
@@ -715,23 +841,30 @@ const key2option = [
       this.showSaveBtn = false
       this.isAble = false
     },
-    /* 提交编辑的内容 */
+    /**
+      * @method：submitEdit
+      * @desc：提交编辑的内容
+      * @params: formName:表单的名称
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     submitEdit(formName) {
       console.log(this.dataForm)
       this.$refs[formName].validate((valid) => {
         if (valid) {
-          // createSoftwareInfo(this.dataForm).then(res => {
-            // this.$message({
-            //   message: '修改成功',
-            //   type: 'success'
-            // })
-            // // 根据返回信息重新复制dataForm
-            // console.log('success submit!!')
-            // // 修改成功后操作
-            // this.afterEdit()
-          // }).catch(err => {
-          //   alert('出错！')
-          // })
+          modifySoftwareInfo(this.dataForm).then(res => {
+            this.$message({
+              message: '修改成功',
+              type: 'success'
+            })
+            // 根据返回信息重新复制dataForm
+            console.log('success submit!!')
+            // 修改成功后操作
+            this.afterEdit()
+          }).catch(err => {
+            alert('出错！')
+          })
         } else {
           this.$message({
             message: '修改内容存在错误，请修改后再保存，或者点击取消按钮取消操作',
@@ -742,13 +875,27 @@ const key2option = [
         }
       })
     },
-    /* 取消编辑操作 */
+    /**
+      * @method：cancelEdit
+      * @desc：取消编辑
+      * @params: formName 表单名称
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     cancelEdit(formName) {
       this.dataForm = { ...this.tempData }
       this.$refs[formName].clearValidate()
       this.afterEdit()
     },
-    /* 取消 确认弹窗 */
+    /**
+      * @method：beforeCancelEdit
+      * @desc：取消编辑，确认弹窗，确认后调用cancelEdit方法
+      * @params: formName 表单名称
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     beforeCancelEdit(formName) {
       this.$confirm('修改信息还没保存, 是否继续?', '提示', {
         confirmButtonText: '确定',
@@ -764,7 +911,14 @@ const key2option = [
         console.log('已取消退出')
       })
     },
-    /* 删除 */
+    /**
+      * @method：handleDelete
+      * @desc：根据ID删除信息
+      * @params:
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     handleDelete() {
       if (this.dataForm.id) {
         deleteSoftwareInfoById(this.dataForm.id).then(res => {
@@ -778,7 +932,14 @@ const key2option = [
         return false
       }
     },
-    /* 删除 确认弹窗 */
+    /**
+      * @method：beforeHandleDelete
+      * @desc：删除前 确认弹窗，确认后调用handleDelete方法
+      * @params:
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     beforeHandleDelete() {
       this.$confirm('此操作将删除该信息, 是否继续?', '提示', {
         confirmButtonText: '确定',
@@ -804,11 +965,25 @@ const key2option = [
         })
       })
     },
-    /* 导出 */
+    /**
+      * @method：handleDownload
+      * @desc：导出表单信息
+      * @params:
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     handleDownload() {
       console.log('导出')
     },
-    /* 跳转到使用说明 */
+    /**
+      * @method：handleAttachment
+      * @desc：跳转到使用说明
+      * @params:
+      * @create date： 2020/7/13
+      * @update date： 2020/7/13
+      * @author：李国烨
+     */
     handleAttachment() {
 
     }
